refactor: reorganize components and surah feature
