+++ app/features/surah/[SurahId]/page.tsx
@@ -1,4 +1,5 @@
 // app/surah/[surahId]/page.tsx
+/* eslint-disable @typescript-eslint/no-explicit-any */
 'use client';
 
 import React, { useEffect, useState, useMemo } from 'react';
@@ -12,7 +13,7 @@ import { useAudio } from '@/app/context/AudioContext';
 
 // --- Interfaces & Data ---
 
-export default function SurahPage({ params }: { params: { surahId: string } }) {
+export default function SurahPage({ params }: any) {
   const [verses, setVerses] = useState<VerseType[]>([]);
   const [translationOptions, setTranslationOptions] = useState<TranslationResource[]>([]);
   const [isLoading, setIsLoading] = useState(true);
