+++ app/page.tsx
@@ -2,5 +2,5 @@
 import { redirect } from 'next/navigation';
 
 export default function HomePage() {
-  redirect('/surah/1');
+  redirect('/features/surah/1');
 }
